change flight precommit packet to block.meta
