+++ src/query/service/src/api/rpc/packets/mod.rs
@@ -15,6 +15,7 @@
 mod packet;
 mod packet_data;
 mod packet_data_precommit;
+mod packet_data_progressinfo;
 mod packet_execute;
 mod packet_executor;
 mod packet_fragment;
@@ -23,8 +24,8 @@ mod packet_publisher;
 pub use packet::Packet;
 pub use packet_data::DataPacket;
 pub use packet_data::FragmentData;
-pub use packet_data::ProgressInfo;
 pub use packet_data_precommit::PrecommitBlock;
+pub use packet_data_progressinfo::ProgressInfo;
 pub use packet_execute::ExecutePartialQueryPacket;
 pub use packet_executor::QueryFragmentsPlanPacket;
 pub use packet_fragment::FragmentPayload;
