+++ src/query/service/src/api/rpc/packets/packet_data.rs
@@ -14,20 +14,17 @@
 
 use std::fmt::Debug;
 use std::fmt::Formatter;
-use std::sync::Arc;
 
 use byteorder::BigEndian;
 use byteorder::ReadBytesExt;
 use byteorder::WriteBytesExt;
 use common_arrow::arrow_format::flight::data::FlightData;
-use common_base::base::ProgressValues;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use tracing::error;
 
+use crate::api::rpc::packets::ProgressInfo;
 use crate::api::PrecommitBlock;
-use crate::sessions::QueryContext;
-use crate::sessions::TableContext;
 
 pub struct FragmentData {
     pub data: FlightData,
@@ -45,14 +42,6 @@ impl Debug for FragmentData {
     }
 }
 
-#[allow(clippy::enum_variant_names)]
-#[derive(Debug)]
-pub enum ProgressInfo {
-    ScanProgress(ProgressValues),
-    WriteProgress(ProgressValues),
-    ResultProgress(ProgressValues),
-}
-
 #[derive(Debug)]
 pub enum DataPacket {
     ErrorCode(ErrorCode),
@@ -62,16 +51,6 @@ pub enum DataPacket {
         progress: Vec<ProgressInfo>,
         precommit: Vec<PrecommitBlock>,
     },
-}
-
-impl ProgressInfo {
-    pub fn inc(&self, ctx: &Arc<QueryContext>) {
-        match self {
-            ProgressInfo::ScanProgress(values) => ctx.get_scan_progress().incr(values),
-            ProgressInfo::WriteProgress(values) => ctx.get_write_progress().incr(values),
-            ProgressInfo::ResultProgress(values) => ctx.get_result_progress().incr(values),
-        };
-    }
 }
 
 impl From<DataPacket> for FlightData {
@@ -100,22 +79,14 @@ impl From<DataPacket> for FlightData {
                     .write_u64::<BigEndian>(precommit.len() as u64)
                     .unwrap();
 
+                // Progress.
+                // TODO(winter): remove unwrap.
                 for progress_info in progress {
-                    let (info_type, values) = match progress_info {
-                        ProgressInfo::ScanProgress(values) => (1_u8, values),
-                        ProgressInfo::WriteProgress(values) => (2_u8, values),
-                        ProgressInfo::ResultProgress(values) => (3_u8, values),
-                    };
-
-                    data_body.write_u8(info_type).unwrap();
-                    data_body
-                        .write_u64::<BigEndian>(values.rows as u64)
-                        .unwrap();
-                    data_body
-                        .write_u64::<BigEndian>(values.bytes as u64)
-                        .unwrap();
+                    progress_info.write(&mut data_body).unwrap();
                 }
 
+                // Pre-commit.
+                // TODO(winter): remove unwrap.
                 for precommit_block in precommit {
                     precommit_block.write(&mut data_body).unwrap();
                 }
@@ -161,23 +132,13 @@ impl TryFrom<FlightData> for DataPacket {
                 let progress_size = bytes.read_u64::<BigEndian>()?;
                 let precommit_size = bytes.read_u64::<BigEndian>()?;
 
+                // Progress.
                 let mut progress_info = Vec::with_capacity(progress_size as usize);
                 for _index in 0..progress_size {
-                    let info_type = bytes.read_u8()?;
-                    let rows = bytes.read_u64::<BigEndian>()? as usize;
-                    let bytes = bytes.read_u64::<BigEndian>()? as usize;
-
-                    progress_info.push(match info_type {
-                        1 => Ok(ProgressInfo::ScanProgress(ProgressValues { rows, bytes })),
-                        2 => Ok(ProgressInfo::WriteProgress(ProgressValues { rows, bytes })),
-                        3 => Ok(ProgressInfo::ResultProgress(ProgressValues { rows, bytes })),
-                        _ => Err(ErrorCode::Unimplemented(format!(
-                            "Unimplemented progress info type, {}",
-                            info_type
-                        ))),
-                    }?);
+                    progress_info.push(ProgressInfo::read(&mut bytes)?);
                 }
 
+                // Pre-commit.
                 let mut precommit = Vec::with_capacity(precommit_size as usize);
                 for _index in 0..precommit_size {
                     precommit.push(PrecommitBlock::read(&mut bytes)?);
